Lokaal Werkende Add patient
De add patient werkt op mijn lokale database. Toen ik m echter de
gegevens van tom's database invoerde werkt het niet. Het probleem is dat
er geen id wordt meegegeven als er een nieuwe patient wordt aangemaakt

+++ HipperEnterprise V0.3/src/main/java/com/service/PatientService.java
@@ -29,11 +29,11 @@ public class PatientService {
         patientDAO.updatePatient(patient);
     }
 
-    public PatientUser getPatient(int id) {
+    public PatientUser getPatient(long id) {
         return patientDAO.getPatient(id);
     }
 
-    public void deletePatient(int id) {
+    public void deletePatient(long id) {
         patientDAO.deletePatient(id);
     }
